+++ src/back/Plants.Application/Requests/GetInstructionRequest.cs
@@ -2,10 +2,10 @@
 
 namespace Plants.Application.Requests;
 
-public record GetInstructionRequest(int Id) : IRequest<GetInstructionResult>;
+public record GetInstructionRequest(long Id) : IRequest<GetInstructionResult>;
 public record GetInstructionResult(bool Exists, GetInstructionResultItem Item);
-public record GetInstructionResultItem(int Id, string Title, string Description,
-    string InstructionText, bool HasCover, int PlantGroupId)
+public record GetInstructionResultItem(long Id, string Title, string Description,
+    string InstructionText, bool HasCover, long PlantGroupId)
 {
     //decoder
     public GetInstructionResultItem() : this (-1, "", "", "", false, -1)
